src/requirements.txt
Django==2.2.9
PyYAML==5.1.2
certifi==2019.9.11
chardet==3.0.4
django-admin-rangefilter==0.5.1
django-autocomplete-light==3.4.1
django-cors-headers==3.1.1
django-debug-toolbar==2.0
django-environ==0.4.5
django-extensions==2.2.5
django-import-export==1.2.0
django-json-widget==0.2.0
django-recaptcha==2.0.5
djangorestframework==3.10.3
drf-yasg==1.17.0
idna==2.8
packaging==19.2
pip==19.3.1
psycopg2-binary==2.8.4
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
sentry-sdk==0.13.0
six==1.12.0
sqlparse==0.3.0
suds-py3==1.3.3.0
ua-parser==0.10.0
urllib3==1.25.6
user-agents==2.1
xlwt==1.3.0
xmltodict==0.12.0
openpyxl==3.0.9
